handleAddSongEventFunction
In this function I build a song object with all the values of the inputs. This occurs on the click event for the add song button. I then call the addSong Function and bring the object to add song as an argument.

ps. I made create Id function to generate an id for the song that does not already exist in the player.