addSongFunction
First, I added the song I created to the player.

Second, I sorted the player so the song would be in the correct place.

Third, I got the new index of the song and used it to place it in the DOM

simultaneously(in the same line of code), while creating the element I place it in the DOM