Event Listener
I added an event listener on the songs div of the Dom.
I set up conditions to see which functions run(if any) dependent on where you click.
I added a remove song function that removes songs from the songs and playlist both in the player and DOM.
I created an Add song function which adds songs to the player and the DOM